+++ frontend/src/components/drawers/WorkspaceDrawer.tsx
@@ -1,5 +1,4 @@
 import {
-	Button,
 	Divider,
 	Drawer,
 	IconButton,
@@ -11,13 +10,9 @@ import {
 } from "@mui/material";
 import { useSelector } from "react-redux";
 import { MouseEventHandler, useState } from "react";
-import { useNavigate } from "react-router-dom";
 import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
 import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
 import WorkspaceListPopover from "../popovers/WorkspaceListPopover";
-import AddIcon from "@mui/icons-material/Add";
-import CreateModal from "../modals/CreateModal";
-import { useCreateDocumentMutation } from "../../hooks/api/workspaceDocument";
 import PeopleIcon from "@mui/icons-material/People";
 import MemberModal from "../modals/MemberModal";
 import { selectWorkspace } from "../../store/workspaceSlice";
@@ -29,15 +24,10 @@ interface WorkspaceDrawerProps {
 
 function WorkspaceDrawer(props: WorkspaceDrawerProps) {
 	const { open } = props;
-	const navigate = useNavigate();
 	const workspaceStore = useSelector(selectWorkspace);
-	const { mutateAsync: createDocument } = useCreateDocumentMutation(
-		workspaceStore.data?.id || ""
-	);
 	const [workspaceListAnchorEl, setWorkspaceListAnchorEl] = useState<
 		(EventTarget & Element) | null
 	>(null);
-	const [createWorkspaceModalOpen, setCreateWorkspaceModalOpen] = useState(false);
 	const [memberModalOpen, setMemberModalOpen] = useState(false);
 
 	const handleOpenWorkspacePopover: MouseEventHandler = (event) => {
@@ -46,16 +36,6 @@ function WorkspaceDrawer(props: WorkspaceDrawerProps) {
 
 	const handleCloseWorkspacePopover = () => {
 		setWorkspaceListAnchorEl(null);
-	};
-
-	const handleCreateWorkspace = async (data: { title: string }) => {
-		const document = await createDocument(data);
-
-		navigate(document.id);
-	};
-
-	const handleCreateWorkspaceModalOpen = () => {
-		setCreateWorkspaceModalOpen((prev) => !prev);
 	};
 
 	const handleMemberModalOpen = () => {
@@ -105,19 +85,6 @@ function WorkspaceDrawer(props: WorkspaceDrawerProps) {
 				</ListItem>
 			</WorkspaceDrawerHeader>
 			<Divider />
-			<ListItem>
-				<Button
-					variant="contained"
-					startIcon={<AddIcon />}
-					sx={{
-						width: 1,
-					}}
-					onClick={handleCreateWorkspaceModalOpen}
-				>
-					New Note
-				</Button>
-			</ListItem>
-			<Divider />
 			<ListItem disablePadding>
 				<ListItemButton onClick={handleMemberModalOpen}>
 					<ListItemIcon>
@@ -126,12 +93,6 @@ function WorkspaceDrawer(props: WorkspaceDrawerProps) {
 					<ListItemText primary="Members" />
 				</ListItemButton>
 			</ListItem>
-			<CreateModal
-				open={createWorkspaceModalOpen}
-				title="Note"
-				onSuccess={handleCreateWorkspace}
-				onClose={handleCreateWorkspaceModalOpen}
-			/>
 			<MemberModal open={memberModalOpen} onClose={handleMemberModalOpen} />
 		</Drawer>
 	);
